+++ mslicer/src/app/mod.rs
@@ -135,7 +135,6 @@ impl App {
 
         let slice_config = self.project.slice_config.clone();
         let mut out = Vec::new();
-        let mut preview_scale = f32::MAX;
 
         let slice_height = slice_config.slice_height.get::<Milimeter>();
         let platform_size = (slice_config.platform_size.xy()).map(|x| x.get::<Milimeter>());
@@ -143,18 +142,13 @@ impl App {
         let platform = slice_config.platform_resolution.cast::<f32>();
         let mm_to_px = platform.component_div(&platform_size).push(1.0);
 
+        // Transform models from world-space to platform-space
         for mesh in meshes.into_iter() {
             let mut mesh = mesh.mesh;
             mesh.set_scale_unchecked(mesh.scale().component_mul(&mm_to_px));
 
-            let (min, max) = mesh.bounds();
-            preview_scale = preview_scale
-                .min(platform_size.x / (max.x - min.x))
-                .min(platform_size.y / (max.y - min.y));
-
-            let pos = mesh.position();
             let offset = (platform / 2.0).push(-slice_height);
-            mesh.set_position_unchecked(pos.component_mul(&mm_to_px) + offset);
+            mesh.set_position_unchecked(mesh.position().component_mul(&mm_to_px) + offset);
             mesh.update_transformation_matrix();
 
             out.push(mesh);
@@ -179,7 +173,7 @@ impl App {
                 file.set_preview(&slice_operation.preview_image());
 
                 let config = slicer.slice_config();
-                slice_operation.add_result(config, (file, voxels), preview_scale);
+                slice_operation.add_result(config, (file, voxels));
             }
         ));
     }